added documentations
added initial documentations for each module in the commom_documentations folder, each py file in the common directory has it's own md file with detailed description about the files and it's utilities.
aslo changed the Readme.md file to link to each of them. it now explains broadly what each of the, sopposed to do and link to to the detailed explnations.
also corrected some typos in file naming